pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ide-devops-mcp"
version = "2.2.1"
description = "IDE DevOps MCP server for interacting with Azure DevOps"
license = {text = "MIT"}
authors = [
    {name = "Microsoft Corporation"}
]
readme = "README.md"
requires-python = ">=3.8"
keywords = ["mcp", "azure-devops", "server"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "mcp>=0.9.0",
    "azure-devops>=7.1.0b4",
    "azure-identity>=1.15.0",
    "azure-core>=1.24.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.8.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0"
]

[project.urls]
Homepage = "https://github.com/microsoft/azure-devops-mcp"
Repository = "https://github.com/microsoft/azure-devops-mcp.git"
Issues = "https://github.com/microsoft/azure-devops-mcp/issues"

[project.scripts]
mcp-server-azuredevops = "azure_devops_mcp.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"